requirements.txt
prometheus_client==0.21.1
psutil==6.1.0
python-dotenv==1.0.1
